fix class Product

+++ src/main/java/Product.java
@@ -4,10 +4,10 @@
 public class Product {
     private String name;
     private int u_id;
-    private float price;
+    private double price;
     private int stock;
 
-    public Product(String name, int u_id, float price, int stock){
+    public Product(String name, int u_id, double price, int stock){
         this.name = name;
         this.u_id = u_id;
         this.price = price;
